Fix pre-commit
